docs: Remove link to custom labware form (#18827)
# Overview

The custom labware form is no longer used but still included in the
manual. Linked text to this form is in Chapter 5, Labware, p. 100.

Instead of this form, the right way to request the fee-based service is
to contact support. Updating text to reflect this new process.

RTC-722

## Test Plan and Hands on Testing

Review for clarity and grammar.

## Changelog

Changes text and removes a link to an older form that's no longer
used/needed.


## Review requests

n/a

## Risk assessment

Low. Small text and link changes.