feat(protocol-designer): cap flow rate at hardware maximum
This PR ensures that the interpolated liquid class flow rate for aspirate, dispense, and blowout does not exceed the range prescribed by the pipette liquid definition's plunger curves. To achieve this, we reuse the max flow rate logic in `getMaxUiFlowRate` in the liquid class form change handler utils, to make sure the populated flow rates from the liquid class are physically possible.

Closes RQA-4151