docs(shared-data): Document the `labwareOffset` property in module definitions (#18863)
This clarifies a field in module definitions.

Surprisingly to me, `labwareOffset` is not an offset from the module's
front-left-bottom corner (specified by `cornerOffsetFromSlot`). It's an
offset from the _parent slot's_ front-left-bottom corner.

To deduce that this is true, look at
`shared-data/module/definitions/3/temperatureModuleV2.json`. Physically,
the module body juts out far beyond the bounds of the slot, but the
module's top slot is basically aligned with the underlying deck slot. In
the definition, this is represented by a big `cornerOffsetFromSlot` x/y
and a small `labwareOffset` x/y.