refactor(app, components): Do not require applying offsets behavior starting a run (#18867)
Closes EXEC-1611

Per product, users are experiencing friction when attempting to start a run due to the requirement that offsets must be formally confirmed by opening LPC and clicking "apply offsets." The solution is to bring LPC more in-line with other setup flows: if offsets haven't been formally applied as they were required in 8.4, we show the "you haven't confirmed XYZ modal", which now includes an entry for applying labware offsets. If the user clicks "confirm", we apply the offsets to the run before proceeding to the run.

Note that the "start run" buttons are still disabled if required default offsets are missing.