fix(app): Fix module positioning in certain deck maps (#18871)
## Overview

Closes EXEC-1645.

## Details

The `<Module>` component knows how to offset itself so it's positioned
correctly relative to the deck, but in order do that, we need to
remember to pass it the `targetDeckId` and `targetSlotId` props, which
are optional. A util feeding a couple of deck maps was forgetting to do
that, so `<Module>` fell back using default, slot-nonspecific
positioning. This was too subtle of an error to be noticeable for most
modules, but it is noticeable for at least the Thermocycler.