docs: Stacker instruction manual (#18664)
Committing Stacker instruction manual files.

# Overview

This is the Stacker instruction manual.

## Test Plan and Hands on Testing

Will review the text for the usual style, grammar, and the usual writing
stuff. Also, check to make sure links work and images appear as
expected. This doesn't have some of the recent CSS, but posting it up
anyway to get started.

## Changelog

No changes. This is new content.

## Review requests

Nothing specific. Please review and comment.

## Risk assessment

Low to medium as this is a new docs feature. Who knows what horrors may
lurk within.

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>