Improve getRunModuleRenderInfo() typing.
These props are nullable when they're passed to <Module> because <Module> needs to support the case where it's being rendered in isolation, not atop a deck. This util, on the other hand, needs to guarantee that it outputs non-null, because it *is* used for aligning modules to decks, and we need these props to compute the transforms to do that.