feat(api): allow lids on h/s to be moved when h/s is latched (#18891)
Helps with EXEC-1647

# Overview

It's been observed that moving lids with a gripper (specifically the
Greiner labware lids) to labware on H/S is more reliable when the
labware itself is latched and doesn't have any room to wiggle on the
H/S- specifically when placed on a universal adapter. So we should allow
lids to be placed on H/S labware with the labware latched.

## Risk assessment

Low- medium. This change will allow any labware categorized as a lid to
be placed on H/S (either directly or on a stack of adpater/labware(s))
without doing a labware latch check. This puts the onus on the user to
make sure to leave the latch closed for the lids. Which is not too much
of a responsibility and has no dire consequences in reality even if the
latch stays open.
We also plan on testing whether we need to add any other restrictions to
improve the user experience.