docs: multi-publication documentation site (#18912)
This PR sets the groundwork for deploying multiple docs publications in
a single site.

# Overview

This PR sets up the
[mkdocs-monorepo-plugin](https://github.com/backstage/mkdocs-monorepo-plugin)
to configure a single mkdocs project that includes smaller mkdocs
projects in their own subdirectories inside of `docs/`.

Publications included in this PR so far:
- Flex manual
- Stacker manual
- HEPA/UV manual

We'll add more publications in future PRs to keep reviews manageable.

## Test Plan and Hands on Testing

There is no CI/sandbox tooling associated with this new setup yet. Test
the docs locally:
- `pip install mkdocs-monorepo-plugin`
- `mkdocs serve` in the top-level `docs/` directory

## Changelog

Lots.

- monorepo docs configuration
- some edits to Stacker manual
- brand new conversion of HEPA/UV manual to mkdocs
- placeholders for PAPI and PD docs

## Review requests

Does this seem functional? If so, let's get it merged and keep building
on it.

## Risk assessment

low-mid. there could be some merge conflicts with open PRs to modify the
Flex manual. if so, i can handle those conflicts post-merge.