fix(protocol-designer): always enable save in EditInstrumentsModal
This PR updates the EditInstrumentsModal component to ensure the "Save" button is always enabled. The behavior of save should be determined by the conditions previously used to set the "disabled" prop on the button. Now, clicking the button when saving if not valid will result in the error message "Add at least one pipette to save" displaying.

Closes AUTH-1321