feat(mono): Support auto-sorting properties in css files (#18937)
Closes AUTH-2109

Sort CSS properties in CSS files by smacss convention and enforce that sorting with lint.

Like the import order sorting, this should be relatively hands off and work on-save in most IDEs, however there is a new format-css makefile command that will sort CSS files according to the plugin rules.