refactor(shared-data): forbid extra fields in liquid class dictionary (#18950)
# Overview

By default, Pydantic silently ignores any extra fields the user provides
when constructing an object. This is bad because liquid classes are a
complex nested model, and it's really easy to put a field in the wrong
place or to make a typo in the field name (like PD was doing).

This change makes Pydantic reject any attempt to construct a liquid
class with fields that are not in the model. After this change, users
will get errors like
```
dispense.retract.disposal_by_volume
  Extra inputs are not permitted [type=extra_forbidden, input_value=[(0, 5)], input_type=list]
```
if they put `disposal_by_volume` in the wrong place.

It would also have helped catch the bug where we accidentally defined
the `multi-dispense` field with a hyphen instead of an underscore
(AUTH-2105), because it would have complained:
```
multi_dispense
  Extra inputs are not permitted [type=extra_forbidden, input_value={'dispense_position': {'o...al_by_volume': [(0, 5)]}, input_type=dict]
```

## Test Plan and Hands on Testing

Let's see if this breaks any existing CI test.

## Risk assessment

Low.