refactor(components): Add `childrenPositioningMode` prop to `<Module>` to prepare for backwards compatibility (#18981)
## Overview

This is something to help make #18890 less risky in the context of our
ongoing release (v8.6.0) and other PRs continually merging into `edge`.

## Changelog

I'm basically making a type-checking tweak that will turn certain merge
mistakes into a hard error. See the new documentation comment for
details.

Without this change, #18890 would need to do a hard cutover where it
updates all uses of `<Module>` at once, which would be error-prone.

This change will be in the v8.6.0 branch. The changes in #18890 will not
be.