refactor(components): Move SVG fudge to <svg> element (#18938)
## Overview

Goes towards EXEC-1278.

## Changelog

The SVG graphic of the Flex Stacker is not visually aligned with the
underlying module geometry. We need to shift it slightly to compensate.

Previously, we did that by fudging the whole `<Module>` component's
position. One disadvantage to that approach is that all the callers of
`<Module>` need to individually remember to special-case the stacker,
and I suspect some of those callers did forget. Another disadvantage is
that it doesn't compose well: the adjustment would apply to `<Labware>`
inside the `<Module>`, but we want it to only apply to the SVG of the
module itself.

To solve both, this moves the adjustment onto the underlying `<svg>`
element.