docs: Add 4-pin DIN connector info to modules (#18884)
# Overview

The Temperature Module and the Heater-Shaker use a 4-pin DIN power
connector that has a unique shape (partly rounded and flat). Sometimes
users don't notice this and try to force the connector into the module.
This can damage the module and connector. Adding a new row to the
specifications tables for each of these modules to show and describe
this connector. It may not prevent this problem, but at least we'll have
documented it.

This is for the Flex manual only. Modules chapter. Reference: RTC-685.
Separate tickets will update quickstarts and instruction manuals.


## Test Plan and Hands on Testing

Check table code and text for accuracy. I've not been using a relative
path to the images file and will start doing that now.

## Changelog

Adds this row and new image to specifications table:

```
<td><strong>Power adapter connector</strong></td>
<td>
   4-pin DIN<br>
   <img src="../images/module-power-connector.png" alt="4-pin DIN connector with flat side" width="40%"><br>
   Align the connector's flat side with the flat side of the module's power port.
</td>
```
Also changes 2 markdown tables to HTML to better handle this complexity.

## Review requests

Putting a border around the image in the cell may make it look odd.
Maybe float it? Not really sure what to do here. Looks OK as is, IMO,
but can change if there's another way. It conveys the idea.

## Risk assessment

Low. Documentation only.