fix(api): post-dispense air gaps in LC based transfers (#19025)
Closes AUTH-2138

# Overview

We were previously using `airGapByVolume` of post-aspirate retract
property for finding air gaps to use during post-dispense retracts. This
PR fixes that so we use the `airGapByVolume` property of a
post-single-dispense or a post-multi-dispense in such cases.

Additionally, when adding air gaps in-between multi-dispenses (when
conditioning volume is zero), we should be using the total tip volume as
the lookup volume for `airGapByVolume` but we were just using `0`uL.

## Risk assessment

Medium. This fixes the air gaps in a lot of cases, but the impact of
this change is not too high practically speaking.