fix(api): do not rely on inexact volume in tip to determine last dispense (#19030)
# Overview

In `distribute_with_liquid_class()`, after we dispense to the last well
of a multi-dispense, we're supposed to blow out the disposal volume.
This is triggered by calling
`retract_during_multi_dispensing(is_last_retract=True)`.

However, we were computing `is_last_retract` as:
```
components_executor.retract_during_multi_dispensing(
    is_last_retract=tip_starting_volume - volume == disposal_volume,
)
```
where `tip_starting_volume` is the current volume in the tip before the
dispense, `volume` is the amount we're about to dispense, and
`disposal_volume` is the disposal volume.

This equality becomes **inexact** when `tip_starting_volume`, `volume`,
and `disposal_volume` are not nice round numbers, which often happens
with liquid classes because the `disposal_volume` is interpolated.

This floating point bug causes `is_last_retract` to be `False`, which
causes us to never dispose of the disposal volume. (AUTH-2151)

This fix changes the code to no longer rely on `tip_starting_volume` to
determine whether we are about to do the last dispense. Instead, in the
loop that performs the multi-dispense, we simply see if we're in the
last iteration of the loop.

## Test Plan and Hands on Testing

I updated the unit tests. Note that we already have existing unit tests
for both the `is_last_retract=False` and `is_last_retract=True` cases.

After my change, I was able to delete
`decoy.when(get_aspirated_volume(...)).then_return(...)` from the unit
tests, which ensures that our implementation no longer looks at the
current volume in the tip.

I also manually analyzed this protocol to check that we are blowing out
the disposal volume:
```
from opentrons import protocol_api, types

requirements = {"robotType": "Flex", "apiLevel": "2.24"}

def run(protocol: protocol_api.ProtocolContext) -> None:
    tip_rack = protocol.load_labware(
        "opentrons_flex_96_tiprack_200ul",
        location="C2",
        namespace="opentrons",
    )
    well_plate = protocol.load_labware(
        "opentrons_96_wellplate_200ul_pcr_full_skirt",
        location="B1",
        namespace="opentrons",
    )
    pipette = protocol.load_instrument(
        "flex_1channel_1000", "left", tip_racks=[tip_rack],
    )
    waste_chute = protocol.load_waste_chute()

    liquid_class = protocol.get_liquid_class(name="water")
    transfer_props = liquid_class.get_for(pipette, tip_rack)
    transfer_props.multi_dispense.retract.blowout.location = "trash"
    # Before this fix, we would not blow out this disposal volume:
    transfer_props.multi_dispense.disposal_by_volume.set_for_all_volumes(1.1)

    pipette.distribute_with_liquid_class(
        liquid_class=liquid_class,
        volume=10,
        source=[well_plate["A1"]],
        dest=[well_plate["G11"], well_plate["H12"]],
        new_tip="once",
        return_tip=True,
    )
```

## Risk assessment

Medium. This is rather complicated code.