docs: update Flex robot and instrument cleaning instructions (#18902)
# Overview

These changes prompted by customer misunderstanding calibration
"recommendations" as "requirements." Trying to correct that in the
cleaning instructions.

Fixing any other code oddities. This one is simple and a low priority on
the list. Getting to the end of the items.

See RTC-647.

## Test Plan and Hands on Testing

Check for code accuracy and grammar.

## Changelog

Text changes to put emphasis on calibration as optional but recommended
when re-attaching an instrument for cleaning. A "nice to have" but not
required tone.

## Review requests

n/a

## Risk assessment

Low, documentation only

---------

Co-authored-by: Ed Cormany <edward.cormany@opentrons.com>