fix(api): failure to create a module instance should be logged and not interrupt hardware controller process (#19010)
# Overview
This pull request improves error handling in
`AttachedModulesControl.register_modules()` by adding a `try-except`
block to ensure that failures during module creation are logged without
interrupting the rest of the process.

This fixes an issue saw in ABR where connecting a module with a new
hardware revision to a robot on an older software, the robot server
would hang.