feat(protocol-designer): add logic for lid move compatibility
This PR extends functionality for moving lids around the deck according to product specs. As part of this work, I extend LabwareTemporalProperties in step-generation and protocol-designer to include a new optional property `isUsedLid`. This property flips to `true` if the lid is loaded on or moved to a "pipettable" labware, which could presumably contain liquid. We check this property's value when determining safe moves to labware stacks. If the lid has been used, we do not allow its further movement to a new pipettable labware for sterility concerns.

Closes AUTH-1071