fix(python): Do math on monotonic clocks only (#19060)
## Overview

A lot of Python code (in various projects) was doing stuff like:

```python
start = time.now()
do_stuff()
end = time.now()
duration = end - start
```

That's not quite right because `time.now()` can be affected by things
like daylight savings time, leap seconds, and NTP adjustments. If we're
doing math on timestamps, it needs to come from a monotonic clock. So
this switches all (or at least most) of those places to use
`time.monotonic()` instead.

## Changelog

Basically a global Ctrl+F for `time.time()`. I replaced it with
`time.monotonic()` when it was obvious to me that it was only used in
the local scope and that math was being done on it. So this excludes
things like `time()` calls that were getting saved in CSV reports.