test(shared-data): check that standard liquid class correction volumes are non-negative (#19075)
# Overview

This adds a check to the validation tests for our standard liquid
classes to make sure we don't have any values that would push the
plunger below the zero position, which would cause a
`PipetteNotReadyToAspirateError`.

Note that this does not address the problem with user-defined liquid
classes. This just helps me be more confident that our standard liquid
classes aren't broken out-of-the-box. AUTH-2154

## Test Plan and Hands on Testing

All of our default liquid classes currently pass.

## Risk assessment

Low, test-only change.