fix(hardware): refactor the runner timeout again
By splitting the timeout on the move group runs into two segments. When
we timeout, we aren't usually actually timing out because the ack took
too long; either it gets or it doesn't. it's just that sometimes we
don't notice that it got there. So what we expect is the first timeout
of minimum 3s will pass unless we don't get scheduled or unless the
message got lost; if it fails, then we want to wait again in case we
didn't get scheduled. And then if the second timeout fails, the message
definitely got lost.