refactor(hardware, shared-data, api): add bounds to unbounded lru caches (#19110)
# Overview

Unbounded caches have the potential for causing memory leaks. From a
code audit standpoint, it is difficult to verify whether or not a
particular cache is responsible for a leak without fulling understanding
how the caching API interacts with its callees and the parameters passed
to the API.

Yes, this is entirely a paranoia PR, but it seems like a reasonable ask
to bound caches so that:

- Any upstream changes to a cached API don't cause memory issues.
Bounding caches eliminates a bug vector.
- Auditing LRU caches in the future is a bit easier.
- Future API authors, when adding a cache to a function, see that the
rest of our codebase always defines a cache size and therefore do so
themselves (or that is hopefully more likely, at least).

What do we think?