Manually fix opentrons_tough_1_reservoir_300ml xDimension/yDimension.
New numbers based on the drawing "SINGLE WELL RESERVOIR SUPPORT DWG" rev A1.0.

xDimension 108->108.7: The old value matched the bottom of the well, but a well's xDimension needs to represent the top of the well. 108.7 matches the drawing and the number in innerLabwareGeometry.

yDimension 72->70.7: I'm not sure where 72 came from, but 70.7 matches the drawing and the number in innerLabwareGeometry.