Manually sync tube rack diameters from the hardware spreadsheet.
We can't easily use a script for this because the spreadsheet doesn't have full information for these definitions--it only has the diameters and depths. Also, some of the tube racks are heterogeneous.

Also, we can only copy over the diameter values, not the depth values. Since the hardware spreadsheet doesn't give us a known tube top position, we can't know whether to grow/shrink the tube depths from the bottom or the top.