Fix diameter in opentrons_6_tuberack_nest_50ml_conical.
The hardware spreadsheet doesn't explicitly mention the diameter of these tubes, so this is a little bit dicey.

These new numbers make the well diameter internally consistent with the definition's innerLabwareGeometry.

A while ago, in PR #17082, this same 27.95 -> 28.18 change was made to opentrons_10_tuberack_nest_4x50ml_6x15ml_conical, which shares the same tubes. I'm guessing this definition was just accidentally omitted.