fix(components): fix title prop null check in DropdownMenu
This PR fixes the logic for nullish `title` prop in `DropdownMenu`, changing the check from `title !== null` to `title != null`, such that an undefined title results in this condition evaluating to `false`. The result is that when `title` is undefined, there will be no 8px spacing above the clickable dropdown box.

Closes RQA-4497