test(api): Catch labware versions not included in any Python Protocol API version (#19146)
## Overview

When you add a new version of an existing labware definition, one big
gotcha is that you need to remember to add it to
`load_labware_params.py`, so new Python Protocol API `apiLevel`s will
load it. If you forget to do this, Python protocols will continue to
load the old labware version. There has never been anything to indicate
that you need to do this, so it's been very easy to accidentally skip.

This PR adds a test that will fail if a new labware version is added
without accounting for it in `load_labware_params.py`.

## Changelog

* Refactor `load_labware_params.py` to be more amenable to testing.
* Add a test that fails if a new labware version is added without being
added to any Python Protocol API `apiLevel`.
* Hard-code a list of temporary exceptions. See the comments for
details.