fix(app,components): Fix unreliable `MoveLabwareOnDeck` animation (#19179)
## Overview

Fixes EXEC-1642 / RQA-4511? Maybe?

## Changelog

Instead of passing in a straight object to `useSpring()`, pass in a
callback and a dependency array. The idea is that maybe we were spamming
react-spring with too many prop changes.

And a completely unrelated thing. The animation has a few steps for
animating the "splash" around the labware. For labware being moved
to/from off-deck, this created a weird (in my opinion) pause where it
looked like nothing was happening, because it was happening off-deck
where you couldn't see it. So, now, conditionally skip those animation
steps when we know they're not going to be visible.