fix(app): allow stacker setup with no pipette (#19160)
We want to allow module setup flows when there aren't pipettes attached,
if the module can be setup without pipettes. Which is some of them, some
of the time, from 3 different places.

Closes RQA-4481

The new behavior applies to stackers and to absorbance readers, since
neither requires a pipette for setup.

The new behavior should be:
- On the desktop app, stacker/reader modules that are not currently in
the deck config should get the setup banner even if no pipette is
attached
- On the desktop app, clicking a stacker's setup banner should launch
you into setup module, which should be completeable all the way through,
even if no pipette is attached
- On the ODD, if either (a) at the time the system boots a
stacker/reader is present and not in the deck config or (b) a
stacker/reader is newly plugged in after boot and not in the deck config
(removing a stacker from the deck config will not trigger this) you
should get the blue new-module-connected banner, even if there's no
pipette
- On the ODD, the module setup wizard should prompt only modules that
both need to be can currently be set up - so, even if there is a
heater/shaker that needs calibration, if you don't have a pipette it
shouldn't appear in the menu, but stackers and readers should always
appear
- When you run out of modules that can currently be calibrated, the flow
should be the same as if you set up all the modules