Also add opentrons_96_aluminumblock_generic_pcr_strip_200ul v4 to apiLevel 2.26.
This was added in a prior PR but never added to an apiLevel, under the assumption that it would be continued to be worked on in this PR, probably to fix up well depths. It didn't actually get changed in this PR because the hardware spreadsheet doesn't give us enough info to know whether the well depths ought to be grown/shrunk from the top or bottom.