fix(shared-data): Sync labware dimensions and spacing to latest hardware values (#19127)
## Overview

Closes EXEC-1480.

## Changelog

* Copies the numbers from the "XY Spacing" tab of the Google Sheet
linked in EXEC-1480.

This means updating the coarse-grained labware and well dimensions, but
*not* the fine-grained `innerWellGeometry`. `innerWellGeometry` should
have already been sorted out by #18266.
* This was partially
[script-assisted](https://github.com/SyntaxColoring/labware-scripts/blob/main/import_dimensions_from_hw.py),
but only partially. The details are in the commit messages.
  
* **Unfortunately, I couldn't bring the tube racks fully up to date**
because the spreadsheet is missing some details. It tells us tube
depths, but it doesn't tell us anything about where the tube sits
vertically. So if we need to fix the depth, we have no way of knowing
whether to grow/shrink it from the top or the bottom.

* Adjust our labware definition tests to only check each labware's
*latest version* for sensible geometry.