refactor(components,app,protocol-designer): Pull labware transforms out of <Module> and prepare them for labware schema 3 (#18890)
## Overview

Another piece towards EXEC-1278. This refactors the way the frontend
renders things aligned to modules (usually labware aligned to modules).
It also updates many of those places so they will work properly when
given a labware in schema 3.

## Changelog

Before, we had a pattern like this:

```jsx
<!-- Renders a module SVG. -->
<Module ...> 
  <!-- Children are rendered at the (-x, -y) corner of the "slot" on top of the module. -->
  
  <!-- Renders a labware SVG in the correct spot atop the module. -->
  <LabwareRender ... />

  <!-- e.g. this could be an "edit labware" button. -->
  <SomeUIControl ... />
</Module>
```

With this PR, that changes to this:

```jsx
<Module childrenPositioningMode="passThrough"...>
  <AlignLabwareToModule ...>
    <LabwareRender ... />
  </AlignLabwareToModule>

  <AlignControlToModule ...>
    <SomeUIControl ... />
  </AlignControlToModule />
</Module>
```

Basically, we're taking an SVG transform that used to be built into
`<Module>` and moving it into its own component. The new helper
components are separated according to their use case:

* `<AlignLabwareToModule>` goes from the module origin to the labware
origin, and it works with labware schema 3. When we implement the
TypeScript code to, for example, align well A1 of a well plate to well
A1 of a Thermocycler, it will automatically benefit from that.
* `<AlignControlToModule>` reimplements the prior behavior: it goes from
the module origin to the (-x, -y) corner of the "slot" on top of the
module. This is intended to support various non-labware things, like
buttons and info overlays, that are rendered in the module's "labware
area."

The old behavior is preserved, for now, if you do `<Module
childrenPositioningMode="offsetToSlot">`. This is basically equivalent
to `<AlignControlToModule>` being built into `<Module>`.

This PR also updates `BaseDeck` to use
`childrenPositioningMode="passThrough"`, which means its API changes
too. Its callers need to use one of the helper components now.