feat(opentrons-ai-server): convert method for token optimization [AUTH-2175] (#19182)
# Overview

Closes AUTH-2175

Currently, `get_relevant_api_docs` method is called every time when we
send a prompt, even if we just say `hi`.
To fix this, we are converting it to a tool where `Anthropic API` will
handle when to call the method.



## Review requests
Go to `Go directly to chat` page and prompt `hi`, it should respond
faster than previous.


## Risk assessment

Low