fix(robot-server): handle deck cal inputs better (#19206)
We had a pretty convoluted way of specifying to fastapi that deck
calibration sessions don't have input parameters. The intent was
definitely that clients would have to omit the createParams key, but
even previously to the change that broke things fastapi was accepting
createParams: {} just fine. But when we upgraded pydantic to 2.11.7,
this stopped being fine: when you sent createParams: {} this would
create an instance of BaseModel, for some abstruse reason relating to
our inheritance hierarchy, which was never something you ought to create
on its own. In 2.11.7, this now eventually causes an error.

The fix is to make sure that the empty-dict input case is explicitly
allowed and tavern-tested - since that's what the app has always done -
and tell fastapi that the createParams argument can be either none or
the instance of an empty subclass of BaseModel, which is fine to
instantiate.

Closes RQA-4519

## testing
- [x] on an ot2, run deck calibration without errors