feat(step-generation): reassign tipracks before `_with_liquid_class` emission
This PR ensures that tipracks are properly assigned to pipettes performing `_with_liquid_class` API methods. Previously, if a pipette was loaded with assignment of tipracks of varying types, the custom liquid class properties may contain the intended tiprack type for that transfer, but the API may attempt to perform the liquid handling with a tiprack of a different type not specified in the liquid class properties (priority order during tiprack assignment). This PR explicitly reassigns the tiprack array directly preceding these liquid handling methods in the event that a pipette has multiple tiprack types assigned.