fix(api, app): add moveStrategy as a store command param so we can recover from a stacker error   (#19178)
# Overview
We currently call `manualRetrieve` when skipping a stacker store labware
step when recovering from an error. This is wrong and would cause the ER
to be stuck in a loop forever.

This PR adds the `moveStrategy` param to our store command so we get the
stacker state update without moving the hardware through the use of
`MANUAL`. This will allow us to properly "skip" the store step in error
recovery.