refactor(api): do stackup to labware origin calculations in labware origin math (#19235)
Works toward EXEC-1690

Whenever we calculate labware origin, what we really care about is the "origin from the lowest deck item to the labware origin". This calculation typically involves offset calculations over multiple deck items and is not sufficiently captured in the notion of "the direct parent of the labware to the labware itself".

All this to say, we can and should encapsulate these notions within labware origin math instead of leaking them elsewhere.